+++ go/l0_battleship/l0_battleship.go
@@ -14,8 +14,8 @@ package l0_battleship
 
 import "meta_puzzles/test"
 
-func getHitProbability(R int, C int, G [][]int) float64 {
-	var ret int = 0
+func getHitProbability(R int32, C int32, G [][]int32) float64 {
+	var ret int32 = 0
 	for _, row := range G {
 		for _, elt := range row {
 			ret += elt
@@ -25,7 +25,7 @@ func getHitProbability(R int, C int, G [][]int) float64 {
 }
 
 type Args struct {
-	G   [][]int
+	G   [][]int32
 	res float64
 }
 
@@ -34,11 +34,11 @@ func (self Args) GetResult() any {
 }
 
 func Tests() uint {
-	wrapper := func(p Args) float64 { return getHitProbability(len(p.G), len(p.G[0]), p.G) }
+	wrapper := func(p Args) float64 { return getHitProbability(int32(len(p.G)), int32(len(p.G[0])), p.G) }
 
 	args_lists := []Args{
-		{[][]int{{0, 0, 1}, {1, 0, 1}}, 0.5},
-		{[][]int{{1, 1}, {1, 1}}, 1.0},
+		{[][]int32{{0, 0, 1}, {1, 0, 1}}, 0.5},
+		{[][]int32{{1, 1}, {1, 1}}, 1.0},
 	}
 
 	return test.RunAllTests("l0_battleship", args_lists, wrapper, 0.000001)
